Add program for December 17th, 2022 (part 1 only)
Part one here was easier than yesterday's to me, but for part 2 I
couldn't quite figure out a way to detect a repeating pattern and where
it occurs. I still think that's the key to solving this, since then you
can get the height at the point it starts repeating, mulitply it, and use
the modulus operator with that # for rocks on the required # of rocks,
and stack the remainder to get the resulting heith. If that makes sense.

Since part 2 asks for 1 trillion I think calculation is the key rather
than raw performance. Sure, the performance of my part 1 could be
improved, but still.