Add program for December 22nd (part1, part2 WiP)
This completes part one, but I ran out of time for part 2. (self-imposed
time limit.)
There's some progress on modeling the cube for part 2, but it took me
long while to reason about how to wrap around based on an arbitrarily
folded out cube. I've got something worked out now, by starting with one
face (the first encountered) and considering it the front, and then from
there traversing the folded open cube and assigning the other sides.

Then since it's a known model we know which face is in each direction if
you view a face head-on. Still needs further implementing of the
Cube::read_map function, and then working out a function in map to do
the wrapping based on the `Cube`.